Reference to add the route label parksmap-backend
The mlbparks application-template only adds parksmap-backend to the service.  It doesn't add it to the route.  So we'd either have to change the application template or add these instructions.  It'd be cleaner to change the application template, but I'm not sure of the impacts to other mlbparks references.  Please let me know if I did something wrong with trying to get it work without adding this label.  This seems to be the only way to get the mlbparks to show up on parksmap.